Revert installing signal handler for SIGTERM and refactor test script
Move trace checker to local_cluster.py
